Make pyre2 64-bit safe
Now compiles on a 64-bit system; previously, complained that you might
have a string size that couldn't fit into an int.  Py_ssize_t is
required instead of plain size_t because Python's is signed.

+++ src/_re2.pxd
@@ -29,7 +29,7 @@ cdef extern from "<map>" namespace "std":
 
 
 cdef extern from "Python.h":
-    int PyObject_AsCharBuffer(object, const_char_ptr *, int *)
+    int PyObject_AsCharBuffer(object, const_char_ptr *, Py_ssize_t *)
     char * PyString_AS_STRING(object)
 
 cdef extern from "stringpiece.h" namespace "re2":
